[doc] test module renamed to tests
